Adds server::stats mod
